backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.10.0
groq==0.11.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9   